pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Suhrob <oopanndaa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.15"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
asyncpg = "^0.24.0"
alembic = "^1.7.4"
tortoise-orm = "^0.17.8"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
aiosqlite = "^0.17.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
